+++ packages/framework/rasengan/src/core/utils/log.ts
@@ -9,13 +9,3 @@ export const logRedirection = (source: string, destination: string) => {
     )}: ${chalk.grey(source)} ${chalk.green('->')} ${chalk.grey(destination)}`
   );
 };
-
-export const logGetRequest = (url: string) => {
-  const time = new Date().toLocaleTimeString();
-
-  console.log(
-    `${chalk.grey(time)} ${chalk.blue('[Rasengan]')} ${chalk.green(
-      'GET'
-    )}: ${chalk.grey(url)}`
-  );
-};